+++ areas/functionapp/src/AzureMcp.FunctionApp/Commands/BaseFunctionAppCommand.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using AzureMcp.Core.Commands;
+using AzureMcp.Core.Commands.Subscription;
+using AzureMcp.Core.Options;
+
+namespace AzureMcp.FunctionApp.Commands;
+
+public abstract class BaseFunctionAppCommand<
+    [DynamicallyAccessedMembers(TrimAnnotations.CommandAnnotations)] TOptions>
+    : SubscriptionCommand<TOptions> where TOptions : SubscriptionOptions, new();